requirements.txt
pandas>=2.0
numpy>=1.24
biopython>=1.83
scipy>=1.11
tqdm>=4.66
pyyaml>=6.0
matplotlib>=3.7
